tomobot/requirements.txt
nextcord
mafic
aiohttp
python-dotenv
wavelink
youtube_dl
PyNaCl
discord.py